+++ src/components/input/Input.jsx
@@ -68,6 +68,10 @@ export default class Input extends Component {
         this.forceUpdate()
     }
 
+    handleChangeMultiple(e: SyntheticInputEvent<any>): void {
+        this.forceUpdate()
+    }
+
     resizeText(){
         this.refs.multiple.style.width = ((this.refs.multiple.value.length + 1) * 8) + 'px'
         this.forceUpdate()
@@ -102,6 +106,8 @@ export default class Input extends Component {
 
         this.setState({
             multipleValue : multiplesVue
+        }, () => {
+            this.forceUpdate()
         })
 
         this.forceUpdate()
@@ -142,12 +148,66 @@ export default class Input extends Component {
         if (onBlur) onBlur(e);
     }
 
+    handleCompositionStart(): void {
+        this.setState({ isComposing: true });
+    }
+
+    handleIconClick(e: SyntheticEvent<any>): void {
+        if (this.props.onIconClick) {
+            this.props.onIconClick(e)
+        }
+    }
+
+    handleKeyUp(e: SyntheticEvent<any>): void {
+        if (this.props.onKeyUp) {
+            this.props.onKeyUp();
+        }
+    }
+
+    handleDownKey(e: SyntheticEvent<any>): void {
+        if (this.props.onKeyDown) {
+            this.props.onKeyDown(e)
+        }
+    }
+
+    handleCompositionEnd(e: SyntheticEvent<any>): void {
+        this.setState({ isComposing: false });
+        this.handleInput(e);
+    }
+
     handleHoveringStart(): void {
+        if (this.props.onMouseEnter) {
+            this.props.onMouseEnter()
+        }
         this.setState({ hovering: true });
     }
 
     handleHoveringEnd(e: SyntheticEvent<any>): void {
+        if (this.props.onMouseLeave) {
+            this.props.onMouseLeave()
+        }
         this.setState({ hovering: false });
+    }
+
+    handleInput(e: SyntheticInputEvent<any>): void {
+        const { isComposing } = this.state;
+        const { onInput } = this.props;
+        if (isComposing) return;
+        if (e.target.value === this.nativeInputValue()) return;
+        if (onInput) onInput(e.target.value);
+        this.nativeInputValue();
+    }
+
+    focus(): void {
+        setTimeout(() => {
+            (this.refs.input || this.refs.textarea).focus();
+        });
+    }
+
+    blur(): void {
+        setTimeout(() => {
+            (this.refs.input || this.refs.textarea).blur();
+        });
     }
 
     setNativeInputValue() {
@@ -310,6 +370,7 @@ export default class Input extends Component {
             clearable,
             id,
             label,
+            value,
             placeholder,
             prefix,
             prefixIcon,
@@ -359,6 +420,7 @@ export default class Input extends Component {
                         style={this.state.textareaStyle}
                         readOnly={readOnly}
                         autoComplete={autoComplete}
+                        onInput={this.handleInput.bind(this)}
                         onChange={this.handleChange.bind(this)}
                         onFocus={this.handleFocus.bind(this)}
                         onBlur={this.handleBlur.bind(this)}
@@ -408,7 +470,7 @@ export default class Input extends Component {
                                             type="text"
                                             onKeyPress={this.resizeText.bind(this)}
                                             onKeyDown={this.handleKeyDownOnMultiple.bind(this)}
-                                            onChange={this.handleChange.bind(this)}
+                                            onChange={this.handleChangeMultiple.bind(this)}
                                             className={this.classNames('cd-select__input')}
                                 />
                             </div>
@@ -421,12 +483,17 @@ export default class Input extends Component {
                               ? (this.state.passwordVisible ? 'text' : 'password')
                               : type}
                         className="cd-input__inner"
+                        value={!Array.isArray(value) ? value : ''}
                         disabled={multiple || this.inputDisabled()}
                         readOnly={readOnly}
                         autoComplete={autoComplete}
+                        onInput={this.handleInput.bind(this)}
+                        onKeyUp={this.handleKeyUp.bind(this)}
+                        onKeyDown={this.handleDownKey.bind(this)}
                         onChange={this.handleChange.bind(this)}
                         onFocus={this.handleFocus.bind(this)}
                         onBlur={this.handleBlur.bind(this)}
+                        onClick={this.handleIconClick.bind(this)}
                         tabIndex={tabindex}
                         aria-label={label}
                         name={label}
@@ -444,7 +511,8 @@ export default class Input extends Component {
                     }
                     {
                         this.getSuffixVisible() &&
-                        <span className="cd-input__suffix">
+                        <span className="cd-input__suffix" onClick={this.handleIconClick.bind(this)}
+                        >
                             <span className="cd-input__suffix-inner">
                                 {
                                     this.showSuffix() &&
@@ -490,7 +558,7 @@ Input.propTypes = {
     autoFocus: PropTypes.bool,
     maxLength: PropTypes.number,
     minLength: PropTypes.number,
-    value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+    value: PropTypes.any,
     clearable: PropTypes.bool,
     showWordLimit: PropTypes.bool,
     validateEvent: PropTypes.bool,
@@ -516,6 +584,8 @@ Input.propTypes = {
     onBlur: PropTypes.func,
     onChange: PropTypes.func,
     onClear: PropTypes.func,
+    onInput: PropTypes.func,
+    onIconClick: PropTypes.func,
 
     // autoComplete
     autoComplete: PropTypes.string,