+++ src/components/input/Input.test.jsx
@@ -10,6 +10,11 @@ describe('Input test', () => {
             <Input placeholder="Input"/>
         );
         expect(w.hasClass('cd-input')).toBeTruthy();
+        w.find('input').simulate('input', {
+            target: {
+                value: 'Teste'
+            }
+        })
     });
 
     it('disabled', () => {
@@ -88,8 +93,10 @@ describe('Input test', () => {
         input.find('input').simulate('focus');
         expect(onFocus.callCount).toBe(1);
 
+        jest.useFakeTimers()
         input.find('input').simulate('blur');
         expect(onBlur.callCount).toBe(1);
+        jest.runAllTimers()
 
         input.find('input').simulate('change', {target: {value: '10'}});
         expect(onChange.callCount).toBe(1);
@@ -218,13 +225,17 @@ describe('Input test', () => {
         input.instance()
 
         expect(input.find('.is-multiple')).toBeTruthy()
-        input.find('.cd-select__input').simulate('change',
-            {
-                target:
-                    {
-                        value: 'Teste'
-                    }
-            }
-        )
+        input.find('.cd-select__input').instance().value = 'Teste'
+        expect(input.instance().getInput().value).toBe('Teste')
+        input.find('.cd-select__input').simulate('change')
+        input.find('.cd-select__input').simulate('keypress')
+        input.find('.cd-select__input').simulate('keydown', {keyCode: 13})
+        expect(input.state().multipleValue).toStrictEqual(['Teste'])
+        input.find('.cd-select__input').simulate('keydown', {keyCode: 8})
+        expect(input.state().multipleValue).toStrictEqual([])
+        input.find('.cd-select__input').instance().value = 'Teste'
+        input.find('.cd-select__input').simulate('keydown', {keyCode: 13})
+        expect(input.state().multipleValue).toStrictEqual(['Teste'])
+        input.find('.cd-tag__close').simulate('click')
     })
 });